requirements.txt
pandas==1.3.5
selenium==4.13.0
webdriver_manager==4.0.2
pytz
flask==2.3.3
convex==0.7.0
python-dotenv==1.0.0